# apps/architectural-service/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "Architectural Service API",
...
    "contact": {
      "name": "DesignSynapse Support",
      "url": "https://designsynapse.com/support",
      "email": "support@designsynapse.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://designsynapse.com/license"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/designs/{design_id}/compliance-checks": {
      "post": {
        "tags": [
          "compliance"
        ],
        "summary": "Request compliance check",
        "description": "Request a building code compliance check for a design",
        "operationId": "request_compliance_check_api_v1_designs__design_id__compliance_checks_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "design_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Design Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ComplianceCheckRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "List compliance checks",
        "description": "List all compliance checks for a design",
        "operationId": "list_compliance_checks_api_v1_designs__design_id__compliance_checks_get",
        "parameters": [
          {
            "name": "design_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Design Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComplianceCheckResponse"
                  },
                  "title": "Response List Compliance Checks Api V1 Designs  Design Id  Compliance Checks Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/compliance-checks/{check_id}": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Get compliance check results",
        "description": "Retrieve compliance check results and status",
        "operationId": "get_compliance_check_api_v1_compliance_checks__check_id__get",
        "parameters": [
          {
            "name": "check_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Check Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ComplianceCheckResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckStatus": {
        "type": "string",
        "enum": [
          "pending",
          "in_progress",
          "completed",
          "failed"
        ],
        "title": "CheckStatus",
        "description": "Status for analysis and check operations."
      },
...
      "ComplianceCheckRequest": {
        "properties": {
          "code_standards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Code Standards",
            "description": "Building code standards to check against (e.g., 'IBC-2021', 'ADA')"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 100
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction",
            "description": "Local jurisdiction for code requirements"
          },
          "check_types": {
            "items": {
              "$ref": "#/components/schemas/ComplianceCheckType"
            },
            "type": "array",
            "title": "Check Types",
            "description": "Types of compliance checks to perform"
          }
        },
        "type": "object",
        "required": [
          "code_standards"
        ],
        "title": "ComplianceCheckRequest",
        "description": "Request schema for compliance check.",
        "example": {
          "check_types": [
            "building_code",
            "accessibility"
          ],
          "code_standards": [
            "IBC-2021",
            "ADA"
          ],
          "jurisdiction": "City and County of San Francisco"
        }
      },
      "ComplianceCheckResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id",
            "description": "Compliance check ID"
          },
          "design_id": {
            "type": "string",
            "format": "uuid",
            "title": "Design Id",
            "description": "Associated design ID"
          },
          "design_version": {
            "type": "string",
            "title": "Design Version",
            "description": "Design version checked"
          },
          "code_standards": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Code Standards",
            "description": "Code standards checked"
          },
          "jurisdiction": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Jurisdiction",
            "description": "Jurisdiction"
          },
          "status": {
            "$ref": "#/components/schemas/CheckStatus",
            "description": "Check status"
          },
          "passed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Passed",
            "description": "Whether all checks passed"
          },
          "violations": {
            "items": {
              "$ref": "#/components/schemas/ViolationDetail"
            },
            "type": "array",
            "title": "Violations",
            "description": "List of violations found"
          },
          "warnings": {
            "items": {
              "$ref": "#/components/schemas/WarningDetail"
            },
            "type": "array",
            "title": "Warnings",
            "description": "List of warnings"
          },
          "recommendations": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Recommendations",
            "description": "General recommendations"
          },
          "report_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Report Url",
            "description": "URL to detailed compliance report"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At",
            "description": "Check start time"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At",
            "description": "Check completion time"
          }
        },
        "type": "object",
        "required": [
          "id",
          "design_id",
          "design_version",
          "code_standards",
          "status",
          "started_at"
        ],
        "title": "ComplianceCheckResponse",
        "description": "Response schema for compliance check."
      },
      "ComplianceCheckType": {
        "type": "string",
        "enum": [
          "building_code",
          "zoning",
          "accessibility",
          "energy",
          "fire_safety",
          "structural",
          "plumbing",
          "electrical",
          "mechanical"
        ],
        "title": "ComplianceCheckType",
        "description": "Types of compliance checks."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "ViolationDetail": {
        "properties": {
          "code_section": {
            "type": "string",
            "title": "Code Section",
            "description": "Code section reference"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Violation description"
          },
          "severity": {
            "type": "string",
            "title": "Severity",
            "description": "Severity level (critical, major, minor)"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location",
            "description": "Location in design where violation occurs"
          },
          "remediation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remediation",
            "description": "Suggested remediation"
          }
        },
        "type": "object",
        "required": [
          "code_section",
          "description",
          "severity"
        ],
        "title": "ViolationDetail",
        "description": "Compliance violation detail."
      },
      "WarningDetail": {
        "properties": {
          "code_section": {
            "type": "string",
            "title": "Code Section",
            "description": "Code section reference"
          },
          "description": {
            "type": "string",
            "title": "Description",
            "description": "Warning description"
          },
          "recommendation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recommendation",
            "description": "Recommended action"
          }
        },
        "type": "object",
        "required": [
          "code_section",
          "description"
        ],
        "title": "WarningDetail",
        "description": "Compliance warning detail."
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}